requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
chatgpt
click==8.1.7
colorama==0.4.6
frozenlist==1.3.3
idna==3.4
importlib-metadata==6.7.0
markdown-it-py==2.2.0
mdurl==0.1.2
multidict==6.0.4
openai==0.27.10
Pillow==9.5.0
Pygments==2.16.1
requests==2.31.0
rich==13.5.2
tls-client==0.2.1
tqdm==4.66.1
typing-extensions==4.7.1
urllib3==1.26.6
yarl==1.9.2
zipp==3.15.0
pytest